requirements.txt
glfw
imgui
opencv-python
pillow
psutil
pycryptodome
PyOpenGL
pywin32
win32con
win32gui
